Connect frontend to backend
